updated compile_targets to work properly

+++ nomake.h
@@ -1,26 +1,27 @@
 #ifndef COMPILATION_IMPLEMENTATION
-int debug_print(char* status, char* msg);
-int debug_print_array(char* status, char** msg);
-int exec(char* args[]);
-int run(char* pathname)
-int run_args(char* pathname[]);
-int len(const char* str1);
+
+unsigned int debug_print(char* status, ...);
+unsigned int debug_print_array(char* status, char** msg);
+unsigned int exec(char* args[]);
+unsigned int run(char* pathname)
+unsigned int run_args(char* pathname[]);
+unsigned int len(const char* str1);
 char* ext(const char* file);
 char* base(const char* path);
-int IS_PATH_DIR(const char* path);
-int IS_PATH_FILE(const char* path);
-int IS_PATH_EXIST(const char* path);
-int MKFILE(const char* file);
-int RMFILE(const char* file);
-int MKDIR(const char* path);
-int RMDIR(const char *path);
-int is_path1_modified_after_path2(Cstr source_path, Cstr binary_path);
-int print_exec(char* args[]);
-int compile_file(const char* file, char* compiler, const char* extension);
-int compile_targets(const char* files[], const char* compiler, const char* extension);
-int compile_dir(char* origin, char* destination, char* compiler, const char* extension);
-#endif
+unsigned int IS_PATH_DIR(const char* path);
+unsigned int IS_PATH_FILE(const char* path);
+unsigned int IS_PATH_EXIST(const char* path);
+unsigned int MKFILE(const char* file);
+unsigned int RMFILE(const char* file);
+unsigned int MKDIR(const char* path);
+unsigned int RMDIR(const char *path);
+unsigned int is_path1_modified_after_path2(const char* source_path, const char* binary_path);
+unsigned int print_exec(char* args[]);
+unsigned int compile_file(char* file, char* destination, char* compiler, const char* extension);
+unsigned int compile_targets(char* files[], char* destination, char* compiler, const char* extension);
+unsigned int compile_dir(char* origin, char* destination, char* compiler, const char* extension);
 
+#endif
 #ifdef COMPILATION_IMPLEMENTATION
 #include <stdio.h>
 #include <stdlib.h>
@@ -51,7 +52,7 @@ int compile_dir(char* origin, char* destination, char* compiler, const char* ext
 #endif // DEBUG
 typedef const char* Cstr;
 
-int debug_print(char* status, ...){
+unsigned int debug_print(char* status, ...){
   va_list args;
   va_start(args, status);
   printf("[%s] ", status);
@@ -61,12 +62,8 @@ int debug_print(char* status, ...){
   return 1;
 }
 
-int debug_print_array(char* status, char** msg){
-  int i;
-  /* 
-  va_list args;
-  va_start(args, status);
-  */
+unsigned int debug_print_array(char* status, char** msg){
+  unsigned int i;
   int sz=sizeof(**msg)/sizeof(msg[0]);
   printf("[%s] ", status);
   for(i=0; i<sz; i++){
@@ -76,7 +73,7 @@ int debug_print_array(char* status, char** msg){
   return 1;
 }
 
-int exec(char* args[]){
+unsigned int exec(char* args[]){
   pid_t id=fork();
   int child_status;
   if(id==0){
@@ -86,11 +83,12 @@ int exec(char* args[]){
     printf("forking failed\n");
   return 0;
   }
+  debug("FILE NAME", args[1]);
   wait(&child_status);
   return 1;
 }
 
-int run_args(char* pathname[]){
+unsigned int run_args(char* pathname[]){
   pid_t id=fork();
   int child_status;
   if(id==0){
@@ -105,7 +103,7 @@ int run_args(char* pathname[]){
   return 1;
 }
 
-int run(char* pathname){
+unsigned int run(char* pathname){
 char* command[]={pathname, NULL};
 if(run_args(command)){
 return 1;
@@ -113,12 +111,12 @@ return 1;
 return 0;
 }
 
-int len(Cstr str1){
+unsigned int len(Cstr str1){
   if(str1==NULL){
     fprintf(stderr, "string was null\n");
     return 0;
   }
-  int i=0;
+  unsigned int i=0;
   while(*str1++){
     i++;
   }
@@ -126,7 +124,7 @@ int len(Cstr str1){
 }
 
 const char* ext(Cstr filename){
-int i;
+unsigned int i;
 int sz=len(filename);
   if(sz<0) return NULL;
   for(i=0; i<sz; i++){
@@ -151,7 +149,7 @@ char* base(Cstr file){
   return retStr;
 }
 
-int IS_PATH_DIR(char* path){
+unsigned int IS_PATH_DIR(char* path){
   struct stat fi;
   if(stat(path, &fi)<0){
     if(errno==ENOENT) 
@@ -166,7 +164,7 @@ int IS_PATH_DIR(char* path){
   return 1;
 }
 
-int IS_PATH_FILE(char* path){
+unsigned int IS_PATH_FILE(char* path){
   struct stat fi;
   if(stat(path, &fi)<0){
     if(errno==ENOENT){
@@ -181,7 +179,7 @@ int IS_PATH_FILE(char* path){
   return 1;
 }
 
-int IS_PATH_EXIST(char* path){
+unsigned int IS_PATH_EXIST(char* path){
   struct stat fi;
   if(stat(path, &fi)<0){
     return 0;
@@ -190,7 +188,7 @@ int IS_PATH_EXIST(char* path){
   return 1;
 }
 
-int MKFILE(char* file){
+unsigned int MKFILE(char* file){
   struct stat fi;
   if(stat(file, &fi)!=0){
     if(creat(file, 0644)<0){
@@ -204,7 +202,7 @@ int MKFILE(char* file){
   return 1;
 }
 
-int RMFILE(char* file){
+unsigned int RMFILE(char* file){
   struct stat fi;
   if(stat(file, &fi)==0){
     if(unlink(file)<0){
@@ -218,7 +216,7 @@ int RMFILE(char* file){
   return 1;
 }
 
-int CLEAN(char* directory, char* extension){
+unsigned int CLEAN(char* directory, char* extension){
 if(directory==NULL || extension==NULL){
     fprintf(stderr, "directory or compiler was null\n");
     return 0;
@@ -244,7 +242,7 @@ if(directory==NULL || extension==NULL){
   return 1;
 }
 
-int MKDIR(char* path){
+unsigned int MKDIR(char* path){
   struct stat fi;
   if(stat(path, &fi)!=0){
     mode_t perms = S_IRWXU | S_IRWXG | S_IRWXO;
@@ -260,9 +258,9 @@ int MKDIR(char* path){
   return 1;
 }
 
-int RMDIR(char *path){
+unsigned int RMDIR(char *path){
   struct stat fi;
-  if(stat(path, &fi)==0){
+  if(stat(path, &fi)==0 && IS_PATH_DIR(path)){
     if(rmdir(path)<0){
       fprintf(stderr, "rmdir error:%s %d\n", path, errno);
       return 0;
@@ -272,7 +270,7 @@ int RMDIR(char *path){
   return 1;
 }
 
-int is_path1_modified_after_path2(Cstr source_path, Cstr binary_path){
+unsigned int is_path1_modified_after_path2(Cstr source_path, Cstr binary_path){
   struct stat fi;
   if(stat(source_path, &fi)!=0){
     fprintf(stderr, "%s doesnt exist\n", source_path); 
@@ -281,16 +279,16 @@ int is_path1_modified_after_path2(Cstr source_path, Cstr binary_path){
   if(stat(binary_path, &fi)!=0){
     fprintf(stderr, "%s doesnt exist\n", binary_path);
   }
-  int binary_time=fi.st_mtime;
+  unsigned int binary_time=fi.st_mtime;
   return source_time>binary_time;
 }
 
-int print_exec(char* args[]){
+unsigned int print_exec(char* args[]){
   printf("executed:%s source:%s binary:%s\n", args[0], args[1], args[3]);
   return 1;
 }
 
-int compile_file(char* file, char* destination, char* compiler, const char* extension){
+unsigned int compile_file(char* file, char* destination, char* compiler, const char* extension){
 if(file==NULL || destination==NULL || compiler==NULL || extension==NULL){
     fprintf(stderr, "origin, destination, compiler or extension was null\n");
     return 0;
@@ -307,45 +305,59 @@ debug_print("COULDNT COMPILE", file);
 return 1;
 }
 
-int compile_targets(char* files[], char* destination, char* compiler, Cstr extension){
+unsigned int compile_targets(unsigned int sz, char* files[], char* destination, char* compiler, Cstr extension){
 if(files==NULL || destination==NULL || compiler==NULL || extension==NULL){
     fprintf(stderr, "file, destination, compiler or extension was null\n");
     return 0;
   }
-  if(strcmp(*files, ".")==0 || strcmp(destination, ".")==0){
-  fprintf(stderr, "file and destination can not be only a dot\n");
+  if(strcmp(*files, ".")==0){
+  fprintf(stderr, "file can not be only a dot\n");
   return 0;
   }
+  unsigned int i;
+  unsigned int j;
   struct stat fi;
-  if(strcmp(ext(*files), extension)==0){
-  int i;
-  int sz=sizeof(**files)/sizeof files[0];
-    for(i=0; i<sz; i++){
-  printf("elems:%d\n", sz); 
-  char* command[]={compiler, files[i], "-o", destination, NULL};
-  if(exec(command)){ 
-  print_exec(command);
-  } else{
-  debug_print_array("COULDNT COMPILE", files);
-  }
+  printf("DEBUG ELEMS:%d\n", sz); 
+  for(i=0; i<sz; i++){
+    if(strcmp(destination, ".")==0){
+      if(strcmp(ext(files[i]), extension)==0){
+  char* command[]={compiler, files[i], "-o", base(files[i]), NULL};
+  printf("FILE:%s BASE:%s\n", command[1], command[3]);
+      if(exec(command)){
+      j++;
+      }
+    } 
     }
-  }
-  else{
-  fprintf(stderr, "%s doesnt have the right extension\n", *files);
-  return 0;
+  if(strcmp(destination, ".")!=0){
+    if(strcmp(ext(files[i]), extension)==0){
+  char buff[PATH_MAX];
+  char* twd=calloc(1, PATH_MAX);
+  strcat(twd, destination); 
+  strcat(twd, "/");
+  strcat(twd, base(files[i]));
+  debug("TWD", twd);
+  char* command[]={compiler, files[i], "-o", twd, NULL};
+	if(exec(command)){
+	j++;
+	}
+      }
+    }
   }
   if(stat(base(*files), &fi)!=0){
   fprintf(stderr, "%s wasnt compiled correctly\n", base(*files));
   return 0;
   }
+  printf("[AMOUNT OF FILES] %d [COMPILED] %d\n", sz, j);
   return 1;
 }
 
-int compile_dir(char* origin, char* destination, char* compiler, Cstr extension){
+unsigned int compile_dir(char* origin, char* destination, char* compiler, Cstr extension){
   if(origin==NULL || destination==NULL || compiler==NULL || extension==NULL){
     fprintf(stderr, "origin, destination, compiler or extension was null\n");
     return 0;
   }
+  unsigned int i=0;
+  unsigned int j=0;
   struct dirent *dirent;
   DIR* source_dir;
   source_dir=opendir(origin);
@@ -353,6 +365,7 @@ int compile_dir(char* origin, char* destination, char* compiler, Cstr extension)
     while((dirent=readdir(source_dir))!=NULL){
       if(strcmp(dirent->d_name, ".")!=0 && strcmp(dirent->d_name, "..")!=0){
 	if(strcmp(ext(dirent->d_name), extension)==0){
+	i++;
 	  char* dest_path=calloc(1, PATH_MAX);
 	  char* origin_path=calloc(1, PATH_MAX);
 	  if(strcmp(origin, destination)==0){
@@ -391,27 +404,28 @@ int compile_dir(char* origin, char* destination, char* compiler, Cstr extension)
 	    debug("BINARY", command[2]);
 	    debug("SOURCE", command[3]);
 	    printf("[source]:%s [binary]:%s\n", origin_path, dest_path);
-	    exec(command);
+	    if(exec(command)){
+	    j++;
+	    }
 	  }
 	}
       }
     }
   }
-  if(strcmp(origin, destination)!=0){
-  printf("finished compiling %s %s\n", origin, destination);
-  }
+  printf("[AMOUNT OF FILES] %d [COMPILED FILES] %d\n", i, j);
   return 1;
 }
 
-#define GO_REBUILD(argc, argv){    						      \
+#define GO_REBUILD(argc, argv, compiler){    					      \
   char* file=__FILE__;      						      	      \
   printf("FILE %s\n", file);     						      \
   assert(file!=NULL && argc>=0); 						      \
   if(is_path1_modified_after_path2(file, argv[0])){				      \
-    char* command[]={"cc", "-o", argv[0], file, NULL};				      \
+    char* command[]={compiler, "-o", argv[0], file, NULL};		              \
+    if(exec(command)){                                                                \
     printf("[COMPILING] %s\n", command[3]);                                           \
-    exec(command); 								      \
-  }										      \
+    }										      \
+  }	                                                                              \
 }
 
 #endif // COMPILATION_IMPLEMENTATION
